core/requirements.txt
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-cron==0.6.0
django-environ==0.11.2
django-recaptcha==4.0.0
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
whitenoise==6.6.0
zope.interface==6.3
